requirements-vllm.txt
torch==2.5.1
torchaudio==2.5.1
torchvision
transformers==4.48.3
accelerate==1.3.0
openai-whisper
onnxruntime-gpu==1.19.0
omegaconf==2.3.0
librosa==0.10.2.post1
sox==1.5.0
modelscope
numpy==1.26.4
six==1.16.0
hyperpyyaml
conformer==0.3.2
diffusers
pillow
sentencepiece
funasr>=1.1.3
protobuf==5.29.3
gradio>=5.16.0
vllm==0.7.2